move the generateToken helper out of the middleware folder and into the file where it's used
